lib: Add host helper
Move functions related to host operation (create temp file,
run command) to new helper. This is needed, because install
engine will use these functions too.
