+++ lib/setupmanagers/virthck/virthck.rb
@@ -3,6 +3,7 @@
 require 'English'
 require './lib/exceptions'
 require './lib/auxiliary/json_helper'
+require './lib/auxiliary/host_helper'
 
 # Virthck class
 class VirtHCK
@@ -95,14 +96,6 @@ class VirtHCK
     nil
   end
 
-  def temp_file
-    file = Tempfile.new('')
-    yield(file)
-  ensure
-    file.close
-    file.unlink
-  end
-
   def run(name, first_time = false)
     sleep(rand(10))
     temp_file do |pid|
@@ -118,28 +111,6 @@ class VirtHCK
     @logger.info('Cleanup host configurations')
     cmd = base_cmd + device_cmd + ['end']
     run_cmd(cmd)
-  end
-
-  def prep_log_stream(stream)
-    stream.strip.lines.map { |line| "\n   -- #{line.rstrip}" }.join
-  end
-
-  def log_stdout_stderr(stdout, stderr)
-    @logger.info('Info dump:' + prep_log_stream(stdout)) unless stdout.empty?
-    return if stderr.empty?
-
-    @logger.warn('Error dump:' + prep_log_stream(stderr))
-  end
-
-  def run_cmd(cmd)
-    temp_file do |stdout|
-      temp_file do |stderr|
-        Process.wait(spawn(cmd.join(' '), out: stdout.path, err: stderr.path))
-        log_stdout_stderr(stdout.read, stderr.read)
-        e_message = "Failed to run: #{cmd.join(' ')}"
-        raise CmdRunError, e_message unless $CHILD_STATUS.exitstatus.zero?
-      end
-    end
   end
 
   def create_client_snapshot(name)
